chore: move `uniqueid.js` from `react-styled-ui/src/shared` to `react-styled-ui/src/utils`
